feat(capi)!: ABI draft 0.9 — status codes everywhere, panic guard, renames, f64 params, generated header
Breaking, by design (backward compat explicitly waived):

- Every fallible function returns MoonlittStatus; silent failures
  abolished (queue overflow -> QUEUE_FULL, bad ids/ranges -> INVALID_ARG
  rejected instead of clamped, consumed handles -> NOT_LOADED).
- ffi_guard! wraps every extern fn: Rust panics become
  MOONLITT_ERR_PANIC + thread-local message instead of unwinding into
  the host process. moonlitt_debug_trigger_panic() lets bindings verify
  the guard end-to-end.
- One error accessor: moonlitt_last_error_message() (thread-local);
  per-handle get_error and session-local accessor removed.
- moonlitt_abi_version() + MOONLITT_ABI_* (draft 0.9.0).
- Naming rule 'family prefix == handle type': mixer_set_*(rt) family ->
  runtime_set_*; set_param_for_track -> runtime_set_track_param;
  start/stop -> start_audio/stop_audio; stop_playback -> stop;
  session_load_file -> session_read_json; session_save ->
  runtime_save_session (final signature, UNSUPPORTED stub until 1.0).
- Backend param values are f64 end-to-end (AudioEvent variants widened;
  Runtime event methods now report ring-buffer overflow as bool); fixes
  silent f64->f32 truncation in node/desktop callers.
- cdylib renamed libmoonlitt (was libmoonlitt_capi; a stale March
  libmoonlitt_ffi.dylib had been silently satisfying the testbed —
  removed).
- cbindgen pipeline: include/moonlitt.h generated + CI drift check.
- C# bindings + FFI testbed rewritten in lockstep; testbed now asserts
  every status (25 no-SF2 / 92 full checks, was 14/49).

+++ crates/moonlitt-core/src/event.rs
@@ -9,7 +9,7 @@ pub enum AudioEvent {
     ProgramChange { channel: u8, program: u8 },
     AllNotesOff,
     SetVolume(f32),
-    SetParam { id: u32, value: f32 },
+    SetParam { id: u32, value: f64 },
     // Mixer control events
     MixerTrackVolume { track_id: u8, volume: f32 },
     MixerTrackPan { track_id: u8, pan: f32 },
@@ -21,10 +21,10 @@ pub enum AudioEvent {
     // Insert effect control
     InsertBypass { track_id: u8, insert_id: u8, bypass: bool },
     // Per-track parameter targeting
-    SetParamForTrack { track_id: u8, param_id: u16, value: f32 },
-    SetInsertParam { track_id: u8, insert_id: u8, param_id: u16, value: f32 },
+    SetParamForTrack { track_id: u8, param_id: u16, value: f64 },
+    SetInsertParam { track_id: u8, insert_id: u8, param_id: u16, value: f64 },
     // Send bus parameter control
-    SetSendBusParam { bus_id: u8, param_id: u16, value: f32 },
+    SetSendBusParam { bus_id: u8, param_id: u16, value: f64 },
     // Track routing (0xFF = master, else = group track ID)
     MixerTrackRoute { track_id: u8, target_id: u8 },
     /// Set external sidechain source for an insert effect.
